Break up attentions.py
