+++ MaxText/layers/gemma2.py
@@ -22,8 +22,7 @@ import jax.numpy as jnp
 
 from flax import linen as nn
 
-from MaxText.common_types import MODEL_MODE_PREFILL, Config
-from MaxText.layers import attentions
+from MaxText.common_types import MODEL_MODE_PREFILL, Config, AttentionType
 from MaxText.layers import quantizations
 from MaxText.layers.attentions import attention_as_linen
 from MaxText.layers.linears import mlp_block
@@ -90,7 +89,7 @@ class Gemma2DecoderLayer(nn.Module):
         float32_logits=cfg.float32_logits,
         quant=self.quant,
         kv_quant=quantizations.configure_kv_quant(cfg),
-        attention_type=attentions.AttentionType.LOCAL_SLIDING,
+        attention_type=AttentionType.LOCAL_SLIDING,
         sliding_window_size=cfg.sliding_window_size,
         attn_logits_soft_cap=cfg.attn_logits_soft_cap,
         model_mode=model_mode,
@@ -188,7 +187,7 @@ class Gemma2DecoderLayer(nn.Module):
         float32_logits=True,
         quant=self.quant,
         kv_quant=quantizations.configure_kv_quant(cfg),
-        attention_type=attentions.AttentionType.GLOBAL,
+        attention_type=AttentionType.GLOBAL,
         attn_logits_soft_cap=cfg.attn_logits_soft_cap,
         model_mode=model_mode,
     )
