+++ MaxText/tests/attention_test.py
@@ -36,10 +36,11 @@ from flax.linen import partitioning as nn_partitioning
 from MaxText import maxtext_utils
 from MaxText import max_utils
 from MaxText import pyconfig
-from MaxText.common_types import DECODING_ACTIVE_SEQUENCE_INDICATOR, MODEL_MODE_AUTOREGRESSIVE, MODEL_MODE_PREFILL, MODEL_MODE_TRAIN
+from MaxText.common_types import DECODING_ACTIVE_SEQUENCE_INDICATOR, MODEL_MODE_AUTOREGRESSIVE, MODEL_MODE_PREFILL, MODEL_MODE_TRAIN, AttentionType
 from MaxText.globals import PKG_DIR
-from MaxText.layers import attentions
-from MaxText.layers.attentions import Attention, MLA, ChunkedCausalMask
+from MaxText.layers.attentions import Attention
+from MaxText.layers.attention_op import ChunkedCausalMask, _make_bidirectional_block_mask, _generate_chunk_attention_mask
+from MaxText.layers.attention_mla import MLA
 
 
 class BidirectionalBlockMaskTest(unittest.TestCase):
@@ -48,7 +49,7 @@ class BidirectionalBlockMaskTest(unittest.TestCase):
   def test_one_block_mask(self):
     bidirectional_mask = np.asarray([[0, 1, 1, 1, 0, 0]])
     # pylint: disable=protected-access
-    block_mask = attentions._make_bidirectional_block_mask(bidirectional_mask)
+    block_mask = _make_bidirectional_block_mask(bidirectional_mask)
     expected_mask = np.asarray(
         [
             [
@@ -66,7 +67,7 @@ class BidirectionalBlockMaskTest(unittest.TestCase):
   def test_two_blocks_mask(self):
     bidirectional_mask = np.asarray([[0, 1, 1, 0, 1, 1]])
     # pylint: disable=protected-access
-    block_mask = attentions._make_bidirectional_block_mask(bidirectional_mask)
+    block_mask = _make_bidirectional_block_mask(bidirectional_mask)
     expected_mask = np.asarray(
         [
             [
@@ -84,7 +85,7 @@ class BidirectionalBlockMaskTest(unittest.TestCase):
   def test_batch_block_masks(self):
     bidirectional_mask = np.asarray([[0, 1, 1, 1, 0, 0], [0, 1, 1, 0, 1, 1]])
     # pylint: disable=protected-access
-    block_mask = attentions._make_bidirectional_block_mask(bidirectional_mask)
+    block_mask = _make_bidirectional_block_mask(bidirectional_mask)
     expected_mask = np.asarray(
         [
             [
@@ -110,7 +111,7 @@ class BidirectionalBlockMaskTest(unittest.TestCase):
   def test_empty_block_mask(self):
     bidirectional_mask = np.asarray([[0, 0, 0, 0, 0, 0]])
     # pylint: disable=protected-access
-    block_mask = attentions._make_bidirectional_block_mask(bidirectional_mask)
+    block_mask = _make_bidirectional_block_mask(bidirectional_mask)
     expected_mask = np.zeros(
         (bidirectional_mask.shape[0], bidirectional_mask.shape[1], bidirectional_mask.shape[1]), dtype=bool
     )
@@ -119,7 +120,7 @@ class BidirectionalBlockMaskTest(unittest.TestCase):
   def test_full_block_mask(self):
     bidirectional_mask = np.asarray([[1, 1, 1, 1, 1, 1]])
     # pylint: disable=protected-access
-    block_mask = attentions._make_bidirectional_block_mask(bidirectional_mask)
+    block_mask = _make_bidirectional_block_mask(bidirectional_mask)
     expected_mask = np.ones(
         (bidirectional_mask.shape[0], bidirectional_mask.shape[1], bidirectional_mask.shape[1]), dtype=bool
     )
@@ -132,7 +133,7 @@ class BidirectionalBlockMaskTest(unittest.TestCase):
     causal_mask = (col_ids <= row_ids)[None, None, None, :, :]
     bidirectional_mask = np.asarray([[0, 1, 1, 1, 0, 0], [0, 1, 1, 0, 1, 1]])
     # pylint: disable=protected-access
-    image_mask = attentions._make_bidirectional_block_mask(bidirectional_mask)
+    image_mask = _make_bidirectional_block_mask(bidirectional_mask)
     combined_mask = causal_mask | image_mask[:, None, None, ...]
     expected_mask = np.asarray(
         [
@@ -190,7 +191,7 @@ class ChunkedCausalMaskTest(unittest.TestCase):
     np.testing.assert_array_equal(actual_mask, expected_mask)
     # Make sure _generate_chunk_attention_mask also produces the same mask
     # pylint: disable=protected-access
-    actual_mask = attentions._generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
+    actual_mask = _generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
     np.testing.assert_array_equal(actual_mask, expected_mask)
 
   def test_full_length_chunk(self):
@@ -206,7 +207,7 @@ class ChunkedCausalMaskTest(unittest.TestCase):
     np.testing.assert_array_equal(actual_mask, expected_mask)
     # Make sure _generate_chunk_attention_mask also produces the same mask
     # pylint: disable=protected-access
-    actual_mask = attentions._generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
+    actual_mask = _generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
     np.testing.assert_array_equal(actual_mask, expected_mask)
 
   def test_single_token_chunk(self):
@@ -222,7 +223,7 @@ class ChunkedCausalMaskTest(unittest.TestCase):
     np.testing.assert_array_equal(actual_mask, expected_mask)
     # Make sure _generate_chunk_attention_mask also produces the same mask
     # pylint: disable=protected-access
-    actual_mask = attentions._generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
+    actual_mask = _generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
     np.testing.assert_array_equal(actual_mask, expected_mask)
 
   def test_non_square_shape(self):
@@ -245,7 +246,7 @@ class ChunkedCausalMaskTest(unittest.TestCase):
     np.testing.assert_array_equal(actual_mask, expected_mask)
     # Make sure _generate_chunk_attention_mask also produces the same mask
     # pylint: disable=protected-access
-    actual_mask = attentions._generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
+    actual_mask = _generate_chunk_attention_mask(mask_shape=mask.shape, chunk_size=chunk_size)
     np.testing.assert_array_equal(actual_mask, expected_mask)
 
   def test_value_error_on_zero_chunk_size(self):
@@ -256,7 +257,7 @@ class ChunkedCausalMaskTest(unittest.TestCase):
       ChunkedCausalMask(shape=(4, 4), chunk_size=-2)
     with self.assertRaises(ValueError):
       # pylint: disable=protected-access
-      attentions._generate_chunk_attention_mask(mask_shape=(4, 4), chunk_size=0)
+      _generate_chunk_attention_mask(mask_shape=(4, 4), chunk_size=0)
 
 
 class AttentionTest(parameterized.TestCase):
@@ -922,7 +923,7 @@ class AttentionTest(parameterized.TestCase):
         attention_kernel="dot_product",
         dtype=self.dtype,
         dropout_rate=self.cfg.dropout_rate,
-        attention_type=attentions.AttentionType.GLOBAL,
+        attention_type=AttentionType.GLOBAL,
         model_mode=MODEL_MODE_TRAIN,
         rngs=self.nnx_rng,
     )
@@ -941,7 +942,7 @@ class AttentionTest(parameterized.TestCase):
         attention_kernel="dot_product",
         dtype=self.dtype,
         dropout_rate=self.cfg.dropout_rate,
-        attention_type=attentions.AttentionType.LOCAL_SLIDING,
+        attention_type=AttentionType.LOCAL_SLIDING,
         sliding_window_size=8,
         model_mode=MODEL_MODE_TRAIN,
         rngs=self.nnx_rng,
@@ -991,7 +992,7 @@ class AttentionTest(parameterized.TestCase):
         attention_kernel="dot_product",
         dtype=self.dtype,
         dropout_rate=self.cfg.dropout_rate,
-        attention_type=attentions.AttentionType.LOCAL_SLIDING,
+        attention_type=AttentionType.LOCAL_SLIDING,
         sliding_window_size=self.max_target_length,
         model_mode=MODEL_MODE_TRAIN,
         rngs=self.nnx_rng,
@@ -1028,7 +1029,7 @@ class MLATest(parameterized.TestCase):
       "enable_checkpointing": False,
       "max_target_length": 128,
       "max_prefill_predict_length": 16,
-      "attention_type": attentions.AttentionType.MLA.value,
+      "attention_type": AttentionType.MLA.value,
       "head_dim": 192,
       "q_lora_rank": 10,
       "kv_lora_rank": 20,
@@ -1234,7 +1235,7 @@ class MLATest(parameterized.TestCase):
         "sa_block_kv_dkv_compute": 128,
         "sa_block_q_dq": 128,
         "sa_block_kv_dq": 128,
-        "attention_type": attentions.AttentionType.MLA.value,
+        "attention_type": AttentionType.MLA.value,
         "q_lora_rank": 4,
         "kv_lora_rank": 4,
         "qk_nope_head_dim": 128,
