+++ MaxText/tests/check_llama4_layers.py
@@ -27,8 +27,9 @@ import unittest
 import jax.numpy as jnp
 from jax.sharding import Mesh
 from jax.experimental import mesh_utils
+
 from MaxText.globals import PKG_DIR
-from MaxText.common_types import MODEL_MODE_TRAIN
+from MaxText.common_types import MODEL_MODE_TRAIN, AttentionType
 from MaxText import pyconfig
 from MaxText import maxtext_utils
 from MaxText.layers import attentions, embeddings, llama4
@@ -38,8 +39,8 @@ Attention = attentions.Attention
 
 # pylint: disable=line-too-long, missing-function-docstring
 
-"""  
-Llama4 Vision RoPE 
+"""
+Llama4 Vision RoPE
 Details https://github.com/huggingface/transformers/blob/main/src/transformers/models/llama4/modeling_llama4.py
 """
 
@@ -653,7 +654,7 @@ class Llama4VisionAttentionTest(unittest.TestCase):
         mesh=self.mesh,
         dropout_rate=0,
         name="self_attention_vision",
-        attention_type=attentions.AttentionType.FULL,
+        attention_type=AttentionType.FULL,
         is_nope_layer=False,
         use_bias_in_projections=True,
         is_vision=True,
